Camera.java is bugged
The camera.java file is what's causing the problem that you can't see your modeling platform, I'd offer a change but I'm more versed in c# and I'm not sure what's wrong with it. It seems like something with the variables, though as I said not sure. I only know that this file is what's causing issues